add test fonctionel
